Guard the News nav link and section against an empty list
Previously the "Events"/"Updates" nav items and section were only
shown when there was content; this PR's merge into "News" dropped that
guard entirely, always rendering the nav link and section even if
there were nothing to show. Move the static items into a shared
app/data/news.ts module so layout.tsx can cheaply check for them
without duplicating page.tsx's async blog/newsletter fetches, and
guard the section itself with the already-computed `news.length > 0`.
