+++ apps/website/app/(home)/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import type { ReactElement, ReactNode } from "react";
 import { Inter } from "next/font/google";
 import { Logo } from "~/components/Logo";
+import { STATIC_NEWS_ITEMS } from "~/data/news";
 import { PostHogProvider } from "../providers";
 import { HomeNavigationMenu } from "./HomeNavigationMenu";
 import "~/globals.css";
@@ -42,7 +43,9 @@ const HomeLayout = async ({
     { href: "/#about", label: "About" },
     { href: "/#plugins", label: "Plugins" },
     { href: "/#resources", label: "Resources" },
-    { href: "/#news", label: "News" },
+    ...(STATIC_NEWS_ITEMS.length > 0
+      ? [{ href: "/#news", label: "News" }]
+      : []),
     { href: "/#talks", label: "Talks" },
     { href: "/#team", label: "Team" },
     { href: "/#supporters", label: "Supporters" },
